+++ kobo/apps/trash_bin/utils/__init__.py
@@ -1,0 +1,10 @@
+from .account import replace_user_with_placeholder
+from .project import delete_asset
+from .trash import move_to_trash, put_back, temporarily_disconnect_signals
+
+__all__ = [
+    'delete_asset',
+    'move_to_trash',
+    'put_back',
+    'replace_user_with_placeholder',
+]
